Add lime.ui.KeyCode, increase size of the keyCode value in C++

+++ project/include/ui/KeyEvent.h
@@ -27,7 +27,7 @@ namespace lime {
 			
 			static void Dispatch (KeyEvent* event);
 			
-			int keyCode;
+			double keyCode;
 			int modifier;
 			KeyEventType type;
